+++ models/GuildProfile.go
@@ -8,7 +8,7 @@ type GuildProfile struct {
 	Kind       string     `json:"kind,omitempty"`
 	CreatedOn  *time.Time `json:"createdOn,omitempty"`
 	Master     *Profile   `json:"master,omitempty"`
-	Population int16      `json:"population,omitempty"`
+	Population uint8      `json:"population,omitempty"`
 	Occupying  string     `json:"occupying,omitempty"`
 	Members    []Profile  `json:"members,omitempty"`
 }
